setup-environment: Generate version.date with real time
Currently SOURCE_DATE_EPOCH is set by default to the date time
of last git commit in Openwrt to make the builds reproducible

Kernel version time in uname and /proc/version is set using
this SOURCE_DATE_EPOCH. Uboot timestamp also depends on the
same mechanism.

To get the actual build date and time in uboot and kernel, we
need to override the default SOURCE_DATE_EPOCH logic by setting
the current date time to a file "version.date" in the buildroot
root dir before running make as required by the script.

Signed-off-by: Ermine Jose <quic_erminejo@quicinc.com>
Change-Id: I2e7e5d57e0005721e2427342fbb20efa5502273d