ipq_open: Enable IPQ_PROVISION Feature by Default
To enable communication with a real AFC server, the Device Under Test (DUT)
must have both the pairing and provisioning packages installed.
During the installation of the pairing package, the following error
is encountered:

open /etc/pairing/pairing_config.json: no such file or directory
Check if registered
panic: runtime error: invalid memory address or nil pointer dereference
[signal SIGSEGV: segmentation violation code=0x1 addr=0x0 pc=0x20fd30]

This error occurs because the "pairing_config.json" file is missing.
This file is generated by the qti-mfg-provision package, which is not
enabled by default in the ipq_open profile.

To resolve this issue and support seamless pairing and provisioning
with the AFC server, enable the IPQ_PROVISION feature by
default in open profiles.

CRs-Fixed: 4175269
Change-Id: I0a947a8ffc6e157bbf339aa4c83951e32db34bfc
Signed-off-by: Priyadarshnee Srinivasan <quic_priys@quicinc.com>